bugfix: can now delete docs if no preview exists

+++ app.rb
@@ -129,7 +129,9 @@ class App < Sinatra::Base
     post('/document/delete/:document_id') do
         if @current_user.logged_in?()
             document = Document.get({document_id: params['document_id'].to_i})
-            FileUtils.remove_file("./public/img/#{document.preview}")
+            if not document.preview.nil?()
+                FileUtils.remove_file("./public/img/#{document.preview}")
+            end
             Document.delete({document_id: params['document_id'].to_i})
         end
         redirect(back)
